improved meshing and fixed bug in latticetransmat

+++ src/tm/Materials/LatticeMaterials/latticetransmat.h
@@ -122,7 +122,7 @@ protected:
     int conType = 0.;
 
     /// Type of conductivity and capcity laws.
-    int capacity = 0.;
+    double capacity = 0.;
 
     /// Relative saturated water content
     double thetaS = 0.;
